refactor: 엔티티가 DTO를 모르도록 수정
