+++ src/main/java/wooteco/subway/domain/member/Favorites.java
@@ -1,14 +1,8 @@
 package wooteco.subway.domain.member;
 
-import wooteco.subway.domain.station.Station;
-import wooteco.subway.service.exception.WrongStationException;
-import wooteco.subway.service.member.dto.FavoriteResponse;
-
 import java.util.HashSet;
-import java.util.List;
 import java.util.Optional;
 import java.util.Set;
-import java.util.stream.Collectors;
 
 public class Favorites {
 
@@ -32,26 +26,6 @@ public class Favorites {
 
     public void remove(Favorite favorite) {
         favorites.remove(favorite);
-    }
-
-    public List<FavoriteResponse> toFavoriteResponses(List<Station> stations) {
-        return favorites.stream()
-                .map(favorite -> toFavoriteResponse(stations, favorite))
-                .collect(Collectors.toList());
-    }
-
-    private FavoriteResponse toFavoriteResponse(List<Station> stations, Favorite favorite) {
-        String sourceName = stations.stream()
-                .filter(station -> station.getId().equals(favorite.getSourceId()))
-                .map(Station::getName)
-                .findFirst()
-                .orElseThrow(WrongStationException::new);
-        String destinationName = stations.stream()
-                .filter(station -> station.getId().equals(favorite.getDestinationId()))
-                .map(Station::getName)
-                .findFirst()
-                .orElseThrow(WrongStationException::new);
-        return new FavoriteResponse(sourceName, destinationName);
     }
 
     public Optional<Favorite> findById(Long sourceId, Long destinationId) {
